requirements.txt
aiohttp==3.11.18
cryptg==0.5.0.post0
hachoir==3.3.0
loguru==0.7.3
pillow==11.2.1
python-dotenv==1.1.0
telethon==1.40.0